requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
openpyxl>=3.1.2
python-docx>=1.1.0
reportlab>=4.0.0
pydantic>=2.0.0
XlsxWriter>=3.0.0
